requirements.txt
demoji==1.1.0
dnspython==2.2.1
envyaml==1.10.211231
fastapi==0.75.0
nltk==3.6.7
numpy==1.22.3
pydantic==1.8.2
pymongo==4.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
spacy==3.2.2
spacy-legacy==3.0.9
spacy-loggers==1.0.1
tweepy==4.7.0
tweet-preprocessor==0.6.0
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.13.4
wasabi==0.9.0
webencodings==0.5.1
wheel==0.37.1
zipp==3.7.0